Created DAO interfaces

+++ src/main/java/com/application/auction/dao/AccountDAO.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AccountDAO extends CrudRepository<Account, Integer> {
+public interface AccountDAO extends CrudRepository<Account, Long> {
 }
