+++ src/main/java/com/application/auction/dao/LotDAO.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LotDAO extends CrudRepository<Lot, Integer> {
+public interface LotDAO extends CrudRepository<Lot, Long> {
 }
